+++ CHANGELOG.md
@@ -13,6 +13,19 @@ The format is based on [Keep a Changelog] and this project adheres to [Semantic 
 
 ---
 
+## [v0.9.0] - 2018/08/11
+
+### Added
+
+- Add `pipenv run exec`
+- Add `post-query-filters`
+
+### Remove
+
+- Remove XmlStrategy.post_query(), use `post-query-filters`
+
+---
+
 ## [v0.8.0] - 2018/05/12
 
 ### Added